feat: fix trailer and truck.

+++ src/main/java/lych/trucks/application/dto/request/TruckRequest.java
@@ -31,13 +31,13 @@ public class TruckRequest implements Serializable {
     private String bodyNumber;
 
     @JsonProperty
-    private float weight;
+    private double weight;
 
     @JsonProperty
-    private float length;
+    private double length;
 
     @JsonProperty
-    private float height;
+    private double height;
 
     @JsonProperty
     private String color;
