+++ src/main/java/lych/trucks/application/dto/response/TruckResponse.java
@@ -30,13 +30,13 @@ public class TruckResponse implements Serializable {
     private String bodyNumber;
 
     @JsonProperty
-    private float weight;
+    private double weight;
 
     @JsonProperty
-    private float length;
+    private double length;
 
     @JsonProperty
-    private float height;
+    private double height;
 
     @JsonProperty
     private String color;
